240. sub category - working with sub category module part 4

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -81,6 +81,10 @@ class CategoryController extends Controller
     public function destroy(Category $category)
     {
         try {
+            if ($category->subCategories()->exists()) {
+                return response()->json(['status' => 'error', 'message' => __('This category has sub categories and can not be deleted')], 402);
+            }
+
             $category->delete();
 
             NotificationService::DELETED();
